+++ crypto/rsa/rsa_ssl.c
@@ -98,7 +98,7 @@ int RSA_padding_add_SSLv23(unsigned char *to, int tlen,
     p += 8;
     *(p++) = '\0';
 
-    memcpy(p, from, (unsigned int)flen);
+    memcpy(p, from, flen);
     return (1);
 }
 
@@ -143,7 +143,7 @@ int RSA_padding_check_SSLv23(unsigned char *to, int tlen,
         RSAerr(RSA_F_RSA_PADDING_CHECK_SSLV23, RSA_R_DATA_TOO_LARGE);
         return (-1);
     }
-    memcpy(to, p, (unsigned int)j);
+    memcpy(to, p, j);
 
     return (j);
 }
