+++ crypto/x509/x509_obj.c
@@ -145,7 +145,7 @@ char *X509_NAME_oneline(X509_NAME *a, char *buf, int len)
         } else
             p = &(buf[lold]);
         *(p++) = '/';
-        memcpy(p, s, (unsigned int)l1);
+        memcpy(p, s, l1);
         p += l1;
         *(p++) = '=';
